+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the extension are documented in this file.
 
 ## [Unreleased]
 
+- Support GCC<10 where .su file entries do not contain full paths.
 ## [0.2.1] - 2022-08-08
 
 - Internal refactoring.